Make icons customizable
