+++ test/widgets/audioControls_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
-import 'package:tingting/ui/audioControls.dart';
+import 'package:tingting/ui/audioControls/audioControls.dart';
 import 'package:tingting/viewModels/tingtingViewModel.dart';
 
 class MockTingTingViewModel extends Mock implements TingTingViewModel {}
